+++ CHANGELOG.md
@@ -3,13 +3,13 @@
 ## Unreleased (0.14.0)
 
 ### 🚨 Breaking changes
-- [](https://github.com/vegaprotocol/vegawallet/pull/) -
+- [507](https://github.com/vegaprotocol/vegawallet/issues/507) - Flag `--level` in service run removed
 
 ### 🗑️  Deprecation
 - [](https://github.com/vegaprotocol/vegawallet/pull/) -
 
 ### 🛠  Improvements
-- [](https://github.com/vegaprotocol/vegawallet/pull/) -
+- [507](https://github.com/vegaprotocol/vegawallet/issues/507) - Redirect logs from command service run to a file
 
 ### 🔥 Removal
 - [](https://github.com/vegaprotocol/vegawallet/pull/) -